+++ CHANGELOG.md
@@ -33,3 +33,7 @@
 - Add the `rental/index` and `rental/image` components
 - Hardcode some data to showcase the new components (this will be populated from somewhere else in the future)
 
+## Part 1.6: Interactive components
+
+- Update the `Rental/Image` component to be interactive
+